moved to MDSEnabledClientService
